+++ src/components/Login/LoginForm.vue
@@ -1,47 +1,23 @@
 <template>
   <div>
-    <div>
-      <v-text-field outlined v-model="emailEntered"> </v-text-field>
-      <div><v-btn class="ma-2" @click="checkEmail()"> Check Email </v-btn></div>
-    </div>
+    <div><v-btn class="ma-2" @click="student"> Student Login </v-btn></div>
+    <div><v-btn class="ma-2" @click="admin"> Admin Login </v-btn></div>
   </div>
 </template>
 
 <script>
-import store from "@/store";
-
 export default {
   name: "LoginForm",
   data() {
-    return {
-      emailEntered: null
-    };
-  },
-  computed: {
-    adminValidation() {
-      return this.$store.state.adminValidation;
-    }
+    return {};
   },
   methods: {
-    async checkEmail() {
-      await store.dispatch("validateAdmin", this.emailEntered);
-      if (this.emailEntered != null && this.emailEntered != "") {
-        if (this.$store.state.adminValidation) {
-          // await this.$store.dispatch("setUser", "admin");
-          // this.$router.push("/home");
-          alert("user has been invited");
-        } else {
-          this.errorMsg =
-            "You can only proceed with admin account creation when it is approved!";
-          setTimeout(() => {
-            this.$store.dispatch("logOut");
-          }, 1000);
-        }
-      } else {
-        alert("please input an email");
-      }
+    student() {
+      this.$router.push("/StudentLogin");
+    },
+    admin() {
+      this.$router.push("/AdminLogin");
     }
-  },
-  async mounted() {}
+  }
 };
 </script>
